all dial is working on dailx
