--- common/netx/network_access.go
@@ -1,1 +1,0 @@
-package netx
